a failed attempt
for each node in tree 1, find the number of nodes with odd distance
and even distance. To get that done, I tried going from node to its
ancestor until we reach root, resulting O(n^2) time to compute that
for the whole tree.

Note: Define root to be 0, I can get the following information in
one DFS, but from there to the info above, I used O(n^2) time

for a given node, find the number of odd distance nodes and even
distance nodes WITHIN the subtree with it as root